Fixxxxx
Ho corretto le viste dei prodotti così che non mostrino sconti pari a 0. I prodotti appena aggiunti avevano sconto 0 anche se il campo veniva lasciato vuoto e comunque uno sconto del genere non è uno sconto. Adesso il controllo è che sconto sia > 0 e funziona come dovrebbe.

Adesso adminAdd, adminEdit e adminDelete reindirizzano alla home se accede un utente o alla login se non si è autenticati.

Ho rimosso admin.php

+++ src/views/adminDelete.php
@@ -4,38 +4,52 @@ use function PRODOTTO\deleteProdotto;
 use function PRODOTTO\getInfoFromProdotto;
 
 session_start();
-if(isset($_SESSION["username"])) {
-    $pagetitle = "Elimina prodotto - Amministrazione";
-    $pagedescription = "Interfaccia per la gestione dei prodotti del negozio";
-    $js = '<script type="text/javascript" src="js/.js"></script>';
-    include "template/header.php";
+if(isset($_SESSION["username"])){
+    if($_SESSION["username"] == "admin") {
+        $pagetitle = "Elimina prodotto - Amministrazione";
+        $pagedescription = "Interfaccia per la gestione dei prodotti del negozio";
+        $js = '<script type="text/javascript" src="js/.js"></script>';
+        include "template/header.php";
 
-    $current_page = "Amministrazione >> Eliminazione prodotto";
-    include "template/breadcrumb.php";
+        $current_page = "Amministrazione >> Eliminazione prodotto";
+        include "template/breadcrumb.php";
 
-    if ($_GET['codArticolo']) {
-        echo '<main id="content">';
+        if ($_GET['codArticolo']) {
+            echo '<main id="content">';
 
-        $info = getInfoFromProdotto($_GET['codArticolo']);
+            $info = getInfoFromProdotto($_GET['codArticolo']);
 
-        if($info !== false) {
-            if(isset($_GET['confirm'])) {
-                if (deleteProdotto($_GET['codArticolo'])) {
-                    echo '<h2>Prodotto eliminato con successo</h2><p>L\'articolo '. $_GET['codArticolo'] .' è stato eliminato con successo</p>';
+            if($info !== false) {
+                if(isset($_GET['confirm'])) {
+                    if (deleteProdotto($_GET['codArticolo'])) {
+                        echo '<h2>Prodotto eliminato con successo</h2><p>L\'articolo '. $_GET['codArticolo'] .' è stato eliminato con successo</p>';
+                    } else {
+                        echo '<h2>Impossibile eliminare il prodotto</h2>';
+                    }
                 } else {
-                    echo '<h2>Impossibile eliminare il prodotto</h2>';
+                    echo '<h2>Eliminazione '. $_GET['codArticolo'] .'</h2>';
+                    echo '<p>Se si è sicuri è possibile ';
+                    echo '<a id="link-conferma" href="adminDelete.php?codArticolo='. $_GET['codArticolo'] .'&confirm=true">procedere con l\'eliminazione dell\'articolo</a></p>';
                 }
             } else {
-                echo '<h2>Eliminazione '. $_GET['codArticolo'] .'</h2>';
-                echo '<p>Se si è sicuri è possibile ';
-                echo '<a id="link-conferma" href="adminDelete.php?codArticolo='. $_GET['codArticolo'] .'&confirm=true">procedere con l\'eliminazione dell\'articolo</a></p>';
+                echo '<h2>404</h2><p class="errore">Prodotto non trovato</p>';
             }
-        } else {
-            echo '<h2>404</h2><p class="errore">Prodotto non trovato</p>';
-        }
 
-        echo '</main>';
+            echo '</main>';
+        }
+        include 'template/footer.php';
+    } else {
+        $host  = $_SERVER['HTTP_HOST'];
+        $uri   = rtrim(dirname($_SERVER['PHP_SELF']), '/\\');
+        $extra = 'home.php';
+        header("Location: http://$host$uri/$extra");
+        exit();
     }
-    include 'template/footer.php';
+} else {
+    $host  = $_SERVER['HTTP_HOST'];
+    $uri   = rtrim(dirname($_SERVER['PHP_SELF']), '/\\');
+    $extra = 'login.php';
+    header("Location: http://$host$uri/$extra");
+    exit();
 }
 ?>
